requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
cachetools==5.3.2
pydantic==2.5.3
pytz==2023.3.post1
